api/requirements.txt
numpy==1.24.3
Flask==3.0.0
Werkzeug>=3.0.0
pymongo==4.6.3
python-dotenv==1.0.0
heyoo==0.1.2
gunicorn